Finalisation!
This is it! Final version of the OmniWheels!
The Subwheels connect firmly and roll well from every angle. The center pieces can be clicked together with ease!
The code is cleaned up and commented, and everything important should be able to be modified on the fly with the variables.
Merging to master!